--- MyList.rb
@@ -1,13 +1,0 @@
-class MyList
-    include MyEnumerable
-    def initialize(*args)
-      @list = args
-    end
-  end
-  
-  def each
-    @list.each do |element|
-      yield(element)
-    end
-  end
-
