+++ my_list.rb
@@ -1,0 +1,10 @@
+class MyList
+  include MyEnumerable
+  def initialize(*args)
+    @list = args
+  end
+end
+
+def each(&block)
+  @list.each(&block)
+end
